api-server/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
blinker==1.8.2
click==8.1.7
Django==4.2.16
django-ninja==1.3.0
djangorestframework==3.15.2
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic-core==2.23.4
sqlparse==0.5.1
typing-extensions==4.12.2
werkzeug==3.0.4
zipp==3.20.2
